Added works and about myself sections for mobile
Hopefully internet problems stopped and this time the commit will go through.

Adding the works part of the mobile version and it's respective CSS.

Adding the About part of the mobile version and it's respective CSS.

Adding icons for the About part and it's respective background image.

Fixed some minor formatting problems I left on yesterdays version (Just to make easier the transition to desktop when time comes, I hope)