Update styles.css
Changed some spacing and/or measures that were being pointed out.

Mostly replaced every % in paddings and margins as I learned those are not, generally speaking, good practice.

Changed the card elements to display as flex, thus easier to manipulate.

Moved the "Get in touch" button from the form to the left, where it should've always been.

Adding a "." before each img url, because that makes the images more compatible*